Move the pure findClearTileNear helper out of App.tsx into mapUtils
A module-scope spawn-position helper (nearest clear MINE_FLOOR tile, used for
placing a dynamically-spawned entrance) was sitting at the top of App.tsx. It
is pure — depends only on mapManager and the tile grid — so it belongs with
the other tile-search helpers in utils/mapUtils.ts. Single call site updated;
tsc- and test-verified.

This is the safe, mechanical part of trimming App.tsx. The bulk (a ~1,550-line
effect-heavy component body and a ~975-line JSX tree) is deliberately left for
an interactive, playtested session: nothing in the test suite exercises App's
runtime behaviour, and main auto-deploys, so a blind split risks shipping a
stale-closure or effect-ordering bug that neither tsc nor the tests would
catch. The seams are known (18 effects group into domain controllers following
the existing use*Controller pattern; the JSX tree has natural sub-tree
boundaries) — that work wants `make dev` open, not an autonomous pass.
